Only write to Delta Lake log when timeline_id matches
On Neon, a table can exist in multiple branches, and allowing them to write to the same Delta Lake log is problematic. This is even worse that a branch can be reset to an earlier point of time. To fix this, we attach Delta Lake log to Neon's timeline, which is unique and only moves forward. Specifically, we persist `neon.timeline_id` when a columnstore table is created and only allow writing to the table's Delta Lake log if the current `neon.timeline_id` matches